+++ src/Connection/Cluster/PredisCluster.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Predis\Connection\Aggregate;
+namespace Predis\Connection\Cluster;
 
 use Predis\Cluster\PredisStrategy;
 use Predis\Cluster\StrategyInterface;