+++ src/Connection/Cluster/RedisCluster.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Predis\Connection\Aggregate;
+namespace Predis\Connection\Cluster;
 
 use Predis\ClientException;
 use Predis\Cluster\RedisStrategy as RedisClusterStrategy;