+++ tests/Predis/Configuration/OptionsTest.php
@@ -27,8 +27,8 @@ class OptionsTest extends PredisTestCase
 
         $this->assertInstanceOf('Predis\Connection\FactoryInterface', $options->connections);
         $this->assertInstanceOf('Predis\Command\FactoryInterface', $options->commands);
-        $this->assertInstanceOf('Predis\Connection\Aggregate\ClusterInterface', $options->cluster);
-        $this->assertInstanceOf('Predis\Connection\Aggregate\ReplicationInterface', $options->replication);
+        $this->assertInstanceOf('Predis\Connection\Cluster\ClusterInterface', $options->cluster);
+        $this->assertInstanceOf('Predis\Connection\Replication\ReplicationInterface', $options->replication);
         $this->assertTrue($options->exceptions);
         $this->assertNull($options->prefix);
     }
@@ -43,16 +43,16 @@ class OptionsTest extends PredisTestCase
             'prefix' => 'prefix:',
             'commands' => $this->getMock('Predis\Command\FactoryInterface'),
             'connections' => $this->getMock('Predis\Connection\FactoryInterface'),
-            'cluster' => $this->getMock('Predis\Connection\Aggregate\ClusterInterface'),
-            'replication' => $this->getMock('Predis\Connection\Aggregate\ReplicationInterface'),
+            'cluster' => $this->getMock('Predis\Connection\Cluster\ClusterInterface'),
+            'replication' => $this->getMock('Predis\Connection\Replication\ReplicationInterface'),
         ));
 
         $this->assertInternalType('bool', $options->exceptions);
         $this->assertInstanceOf('Predis\Command\FactoryInterface', $options->commands);
         $this->assertInstanceOf('Predis\Command\Processor\ProcessorInterface', $options->prefix);
         $this->assertInstanceOf('Predis\Connection\FactoryInterface', $options->connections);
-        $this->assertInstanceOf('Predis\Connection\Aggregate\ClusterInterface', $options->cluster);
-        $this->assertInstanceOf('Predis\Connection\Aggregate\ReplicationInterface', $options->replication);
+        $this->assertInstanceOf('Predis\Connection\Cluster\ClusterInterface', $options->cluster);
+        $this->assertInstanceOf('Predis\Connection\Replication\ReplicationInterface', $options->replication);
     }
 
     /**