+++ tests/Predis/Pipeline/FireAndForgetTest.php
@@ -42,7 +42,7 @@ class FireAndForgetTest extends PredisTestCase
      */
     public function testSwitchesToMasterWithReplicationConnection()
     {
-        $connection = $this->getMock('Predis\Connection\Aggregate\ReplicationInterface');
+        $connection = $this->getMock('Predis\Connection\Replication\ReplicationInterface');
         $connection->expects($this->once())
                    ->method('switchTo')
                    ->with('master');